add support for dosage rates
some customers send us infusion rates based on a patient's bodyweight.
for example `10 mcg/kg/hr`. while we're not interested in actually
calculating the mass of the drug administered at this point, we do
need to be able to recognize this as a rate. for now, this commit adds
support to simply drop the `kg` from this rate and pass the rate along
as `10 mcg/hr`. This isn't technically correct, but it serves our
purposes for now.

knows rates that are covered as part of this:
- `mcg/kg/min`
- `mcg/kg/hr`
- `mg/kg/hr`